refactor(project): extract pure topology helpers from api.go
internal/project/api.go is mostly `gh api graphql ...` shell-outs
(17 of 19 functions), which are network/CLI bound and not unit-test
material. Two pure functions — DetectTopology and ControlPlaneRepo —
were stranded inside the same file. They have real test coverage
(api_test.go) but the file's per-file coverage was 2.9%, dragging
the new-code coverage gate below 80%.

Move the two pure functions to a new topology_detect.go (separate
from the existing topology.go, which holds the unrelated and more
elaborate ResolveTopology). Tests in api_test.go continue to pass
unchanged — Go test files address the package, not the source file.

Then add internal/project/api.go to sonar.coverage.exclusions, in
the same category as the auth helpers and mount/releases.go that
were excluded in the previous cleanup pass. The pure functions
remain covered (now in topology_detect.go); only the network
wrappers are off the meter.

Expected effect on the new-code coverage gate: ~71.8% → comfortably
above 80%.
