+++ sonar-project.properties
@@ -37,6 +37,7 @@ sonar.coverage.exclusions=\
   internal/testutil/**,\
   internal/mount/mounttest/**,\
   internal/mount/releases.go,\
+  internal/project/api.go,\
   internal/projectstatus/queries_default.go,\
   internal/auth/credentials.go,\
   internal/auth/owner.go,\
